fsck.s3gw: skeleton and first check
Bring up a skeleton for the code repository and add the first check:
orphaned objects check.
This check ensures that all objects foun in the directory structure are
referenced in the metadata. The fix for any objects that aren't found in
the metadata is to move them to a lost+found location where they can be
retrieved from by an adminstrator.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>