checks: add metadata version check
Checks:

- Add metadata version check

This check is a bit special, because it does not have a fix. It is
intended to be used when doing a preliminary check if the volume could
be re-used for running the s3gw without actually running the s3gw,
because the s3gw aggressively upgrades the metadata schema.
This check doesn't have a corresponding fix, as the fix would be to just
run the newer s3gw and let it upgrade the metadata if that is possible.

- Fix the fixer for the orphaned objects to correcly move/remove objects
  and directories if an orphaned object is in the same directory as a
  non-orphaned object.

Build infra:

- Move CMakeList.txt to source directory

This primarily makes it easier to build in Docker as only the source
directory, not the entire git tree has to be copied into the container.

- Add Dockerfile

This is intended, so the fsck.s3gw tool can be built and distributed as
container image as well.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>