Add ObjectIntegrityCheck
So far this just checks object size, as the checksum column in the database
seems to be empty, so there's nothing to verify there yet.

Conceptually this *is* a separate check to the orphaned metadata check, but
the implementation of the two is almost identical, which begs the question:
Why not just combine the two.  Is the Check abstraction really helping here?

Also, ObjectIntegrityFix has a free text "reason" member.  Adding this was
a somewhat arbitrary decision on my part, but I wanted more information to
be displayed about what was wrong, and it's the Fix class that returns text,
not the Check class.  This relationship probably wants re-examining too.

Signed-off-by: Tim Serong <tserong@suse.com>